// src/renderer/interfaces/states.tsx
export interface InventoryFilters {
  inventoryFilter: Array<Filter>;
  storageFilter: Array<Filter>;
  sortValue: string;
  inventoryFiltered: Array<ItemRow>;
  storageFiltered: Array<ItemRow>;
  searchInput: string;
  sortBack: boolean;
  categoryFilter: Array<string>;
  rarityFilter: Array<string>;
}